+++ ngsupport/ngextras/segmentcolors.py
