+++ ngsupport/ngextras/segmentprops.py
