+++ ngsupport/ngextras/util.py
